[refactor]: 패키지 구조 도메인 단위로 리팩토링
