--- src/main/java/com/ggamja/domain/Card.java
@@ -1,41 +1,0 @@
-package com.ggamja.domain;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Table(
-        name = "card",
-        uniqueConstraints = {
-                @UniqueConstraint(columnNames = {"category", "title"})
-        }
-)
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Card {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private Category category;
-
-    @Column(nullable = false, length = 255)
-    private String title;
-
-    @Column(nullable = false, length = 255)
-    private String meaning;
-
-    @Column(nullable = false)
-    private int difficulty;   // 난이도 (1 ~ 3)
-
-    @OneToMany(mappedBy = "card", cascade = CascadeType.ALL, orphanRemoval = true)
-    private List<Quiz> quizzes = new ArrayList<>();
-}
