--- src/main/java/com/ggamja/domain/Category.java
@@ -1,7 +1,0 @@
-package com.ggamja.domain;
-
-public enum Category {
-    수도,
-    속담
-    // 필요에 따라 더 추가 가능
-}
