--- src/main/java/com/ggamja/domain/Quiz.java
@@ -1,37 +1,0 @@
-package com.ggamja.domain;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Table(name = "quiz")
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Quiz {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, length = 255)
-    private String question;
-
-    @Column(nullable = false, length = 255)
-    private String option1;
-
-    @Column(nullable = false, length = 255)
-    private String option2;
-
-    @Column(nullable = false, length = 255)
-    private String option3;
-
-    @Column(nullable = false, length = 255)
-    private String answer;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "card_id", nullable = false)
-    private Card card;
-}
-
