--- src/main/java/com/ggamja/domain/TodayCard.java
@@ -1,25 +1,0 @@
-package com.ggamja.domain;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.LocalDate;
-
-@Entity
-@Table(name = "today_card")
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class TodayCard {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "card_id", nullable = false)
-    private Card card;
-
-    @Column(nullable = false)
-    private LocalDate date;
-}
