+++ src/main/java/com/ggamja/domain/card/entity/Category.java
@@ -1,0 +1,7 @@
+package com.ggamja.domain.card.entity;
+
+public enum Category {
+    수도,
+    속담
+    // 필요에 따라 더 추가 가능
+}
