+++ src/main/java/com/ggamja/domain/member/dto/request/PostMemberLoginRequest.java
@@ -1,0 +1,13 @@
+package com.ggamja.domain.member.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Email;
+
+public record PostMemberLoginRequest(
+        @NotBlank
+        @Email(message = "이메일 형식이 올바르지 않습니다.")
+        String email,
+
+        @NotBlank
+        String password
+) {}
