+++ src/main/java/com/ggamja/domain/member/dto/response/PostMemberRegisterResponse.java
@@ -1,0 +1,6 @@
+package com.ggamja.domain.member.dto.response;
+
+public record PostMemberRegisterResponse(
+        Long id,
+        String nickname
+) {}
