+++ src/main/java/com/ggamja/domain/quizlog/entity/QuizLog.java
@@ -1,0 +1,40 @@
+package com.ggamja.domain.quizlog.entity;
+
+import com.ggamja.domain.member.entity.Member;
+import com.ggamja.domain.quiz.entity.Quiz;
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Table(name = "quiz_log")
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class QuizLog {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id", nullable = false)
+    private Member member;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "quiz_id", nullable = false)
+    private Quiz quiz;
+
+    @Column(nullable = false, length = 255)
+    private String selectedAnswer;   // 사용자가 선택한 정답
+
+    @Column(nullable = false, updatable = false)
+    private LocalDateTime date;
+
+    @PrePersist
+    protected void onCreate() {
+        this.date = LocalDateTime.now();
+    }
+}
