--- src/main/java/com/ggamja/dto/response/PostMemberLoginResponse.java
@@ -1,21 +1,0 @@
-package com.ggamja.dto.response;
-
-import com.ggamja.domain.Member;
-
-public record PostMemberLoginResponse(
-        Long id,
-        String nickname,
-        String levelName,
-        boolean bonusGiven,
-        boolean levelChanged
-) {
-    public static PostMemberLoginResponse of(Member member, boolean bonusGiven, boolean levelChanged) {
-        return new PostMemberLoginResponse(
-                member.getId(),
-                member.getNickname(),
-                member.getLevel().getLevel(),
-                bonusGiven,
-                levelChanged
-        );
-    }
-}
