--- src/main/java/com/ggamja/dto/response/PostMemberRegisterResponse.java
@@ -1,6 +1,0 @@
-package com.ggamja.dto.response;
-
-public record PostMemberRegisterResponse(
-        Long id,
-        String nickname
-) {}
