--- src/main/java/com/ggamja/exception/MemberException.java
@@ -1,15 +1,0 @@
-package com.ggamja.exception;
-
-public class MemberException extends RuntimeException {
-    private final MemberExceptionResponseStatus status;
-
-    public MemberException(MemberExceptionResponseStatus status) {
-        super(status.getMessage());
-        this.status = status;
-    }
-
-    public MemberExceptionResponseStatus getStatus() {
-        return status;
-    }
-}
-
